try (and give up) making the profile name input slide nicely
what I want is a horizontal accordion-like (or "carousel") effect, where the
<span> slides out to the left while the <input> slides in from the right.
What I've got is close, but when both elements are visible at the same time,
they stack vertically (as if one was a <div>). So the whole "Name" field
doubles in height when the effect starts, and then goes back to normal when
it finishes: very jarring. I suspect there's a way to style this so that the
two elements are willing to share the horizontal space, but I haven't found
it yet.