get drop-profile-icon working, barely
still wants better styling (icon should be centered in the drop-target,
drop-target should sit to the right of the "Icon:" LI, helper text at bottom
should be centered under the drop-target).

Also, it wants to be implemented with jquery's .droppable() code, instead of
manual event handlers for 'drop' and 'dragenter' and 'dragover'. Also need to
update cursor to indicate we're in the drop target.