make bin/tool use the venv python, which ignores system site-packages
We could instead use the system python and add the venv's site-packages
to sys.path, but it turns out that zope.interface uses a "namespace"
package, and has a .pth file in the venv's site-packages to help, and a
simple sys.path.insert(0, venvsitepackages) doesn't process the .pth
file. This results in twisted failing to import because the new twisted
needs a newer zope.interface than what the OS provides.

So we just use venv/bin/python.